Split App into pages
Create an Overview page component that loads the metadata and displays a
list of test folders.

Move previous App code to Slides page component and add require campaign
prop to load different campaigns. Pass campaign prop down to children to
create image paths.
